[Refactor] 스토리북에서 헤더 컴포넌트 구조 변경
