--- src/components/Header/SigInSingUp.stories.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-import { SigninButton } from './SigninButton';
-
-export default {
-  title: 'sign_in',
-  component: SigninButton,
-};
-
-export const Default = (): JSX.Element => <SigninButton />;
